Created new refuse feedback form and moved logic over
