+++ app/forms/referee_interface/refuse_feedback_form.rb
@@ -6,25 +6,8 @@ module RefereeInterface
 
     validates :choice, presence: true
 
-    def save(reference)
-      return false unless valid?
-
-      reference.update!(feedback_status: :feedback_refused, feedback_refused_at: Time.zone.now)
-      send_slack_notification(reference)
-      SendNewRefereeRequestEmail.call(reference: reference, reason: :refused)
-    end
-
     def referee_has_confirmed_they_wont_a_reference?
       choice == 'yes'
-    end
-
-  private
-
-    def send_slack_notification(reference)
-      message = ":sadparrot: A referee declined to give feedback for #{reference.application_form.first_name}’s application"
-      url = Rails.application.routes.url_helpers.support_interface_application_form_url(reference.application_form)
-
-      SlackNotificationWorker.perform_async(message, url)
     end
   end
 end
